backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
openai
google-generativeai
plyer
python-dotenv
youtube-transcript-api
